Revert "Wip: remove centos, rhel, suse, photonos, trusty"
This reverts commit 6acd0c9cf71edd6fab16b094e0d1083cbbd5ff5a.

NOTE: There were MANY merge conflicts which needed to be manually resolved. See notes below for details of the commits which were used to help resolve individual files' conflicts.

# Conflicts:
#	README.md
#	bosh-stemcell/spec/bosh/stemcell/definition_spec.rb
#	bosh-stemcell/spec/bosh/stemcell/operating_system_spec.rb
#	bosh-stemcell/spec/bosh/stemcell/stage_collection_spec.rb
#	stemcell_builder/lib/prelude_agent.bash
#	stemcell_builder/lib/prelude_apply.bash
#	stemcell_builder/stages/dev_tools_config/assets/generate_dev_tools_file_list_ubuntu.sh
#	stemcell_builder/stages/image_install_grub/apply.sh
#	stemcell_builder/stages/logrotate_config/apply.sh
#	stemcell_builder/stages/password_policies/apply.sh
#	stemcell_builder/stages/rsyslog_config/apply.sh
#	stemcell_builder/stages/system_google_packages/apply.sh
#	stemcell_builder/stages/system_google_packages/config.sh
#	stemcell_builder/stages/system_grub/apply.sh

Merge conflict notes:
- commits bccc925d, 0fdf5612 RE: stemcell_builder/stages/system_google_packages/apply.sh (Is python install still needed for rhel|centos? Probably not.)
- commits 8e6cabe4, bbd37202 RE: stemcell_builder/stages/system_google_packages/config.sh (should this file NOT be re-added as part of this REVERT?)
- commits 6acd0c9c, 17339ac8: RE: stemcell_builder/stages/dev_tools_config/assets/generate_dev_tools_file_list_ubuntu.sh (should this previously-renamed file and the other OS-suffixed files be merged?)
- commits c8b82d91, 2108b483: RE: bosh-stemcell/spec/bosh/stemcell/stage_collection_spec.rb (Do other chanes need to be made outside this file, since the 'system_ixgbevf' stage was deleted and the 'clean_machine_id' stage was added for Ubunutu?)
- commit 3c794ed1: RE: bosh-stemcell/spec/bosh/stemcell/definition_spec.rb (I manually re-added some of the specs removed by 3c794ed1. Since multi-OS is supported again, re-adding the specs make sense.)
- commit 3c794ed1: RE: bosh-stemcell/spec/bosh/stemcell/operating_system_spec.rb (FYI: I added a new spec for RHEL OS which was not present in either branch's history.)
- commit dbd3bf9e: RE: stemcell_builder/stages/image_install_grub/apply.sh (dbd3bf9e made major changes to this file which conflict with the re-added code from reverting 6acd0c9c.)
- commit dbd3bf9e: RE: stemcell_builder/stages/image_install_grub/apply.sh (Fixed a comment that wasn't updated in dbd3bf9e: Reintroduce grub2 configuration for bionic)
- commit bbae678f: RE: stemcell_builder/stages/image_install_grub/apply.sh (Manually resolved the conflicts.)

+++ stemcell_builder/stages/dev_tools_config/apply.sh
@@ -6,6 +6,10 @@ base_dir=$(readlink -nf $(dirname $0)/../..)
 source $base_dir/lib/prelude_apply.bash
 source $base_dir/lib/prelude_bosh.bash
 
-cp -p $assets_dir/generate_dev_tools_file_list.sh $chroot/bin/generate_dev_tools_file_list
-run_in_chroot $chroot "/bin/generate_dev_tools_file_list | sort | uniq > /var/vcap/bosh/etc/dev_tools_file_list"
-rm $chroot/bin/generate_dev_tools_file_list
+os_type=$(get_os_type)
+
+if [ "${os_type}" == "centos" ] || [ "${os_type}" == "ubuntu" ] || [ "${os_type}" == "opensuse" ] ;  then
+  cp -p $assets_dir/generate_dev_tools_file_list_$os_type.sh $chroot/bin/generate_dev_tools_file_list
+  run_in_chroot $chroot "/bin/generate_dev_tools_file_list $OS_TYPE | sort | uniq > /var/vcap/bosh/etc/dev_tools_file_list"
+  rm $chroot/bin/generate_dev_tools_file_list
+fi
